# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HRMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      parameters:
        - name: search
          in: query
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: Vendor list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"

  # ── VISITS ────────────────────────────────────────────
  /visits:
    get:
      operationId: listVisits
      tags: [visits]
      parameters:
        - name: employeeId
          in: query
          schema: { type: ["integer", "null"] }
        - name: vendorId
          in: query
          schema: { type: ["integer", "null"] }
        - name: date
          in: query
          schema: { type: ["string", "null"] }
        - name: status
          in: query
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: Visit records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visit"
    post:
      operationId: createVisit
      tags: [visits]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}:
    get:
      operationId: getVisit
      tags: [visits]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/summary:
    get:
      operationId: getVisitSummary
      tags: [visits]
      parameters:
        - name: employeeId
          in: query
          schema: { type: ["integer", "null"] }
        - name: date
          in: query
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: Daily visit summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitSummary"

  # ── PAYROLL ───────────────────────────────────────────
...
components:
  schemas:
...
    Vendor:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        contactPerson: { type: ["string", "null"] }
        mobile: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
        radius: { type: ["integer", "null"] }
        visitCount: { type: ["integer", "null"] }
        lastVisit: { type: ["string", "null"] }
        createdAt: { type: string }
    VendorInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        contactPerson: { type: ["string", "null"] }
        mobile: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
        radius: { type: ["integer", "null"] }
    VendorUpdate:
      type: object
      properties:
        name: { type: ["string", "null"] }
        contactPerson: { type: ["string", "null"] }
        mobile: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
        radius: { type: ["integer", "null"] }

    # ── VISIT ─────────────────────────────────────────
    Visit:
      type: object
      required: [id, employeeId, vendorId, visitDate, status]
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        employeeName: { type: ["string", "null"] }
        vendorId: { type: integer }
        vendorName: { type: ["string", "null"] }
        visitDate: { type: string }
        checkInTime: { type: ["string", "null"] }
        selfieUrl: { type: ["string", "null"] }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
        remarks: { type: ["string", "null"] }
        meetingNotes: { type: ["string", "null"] }
        orderValue: { type: ["number", "null"] }
        nextFollowUp: { type: ["string", "null"] }
        status: { type: string }
        invalidReason: { type: ["string", "null"] }
        createdAt: { type: string }
    VisitInput:
      type: object
      required: [employeeId, vendorId, selfieUrl]
      properties:
        employeeId: { type: integer }
        vendorId: { type: integer }
        selfieUrl: { type: string }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
        remarks: { type: ["string", "null"] }
        meetingNotes: { type: ["string", "null"] }
        orderValue: { type: ["number", "null"] }
        nextFollowUp: { type: ["string", "null"] }
    VisitSummary:
      type: object
      required: [date, totalVisits, validVisits, invalidVisits, totalKm]
      properties:
        date: { type: string }
        employeeId: { type: ["integer", "null"] }
        totalVisits: { type: integer }
        validVisits: { type: integer }
        invalidVisits: { type: integer }
        totalKm: { type: number }
        workingHours: { type: number }
        ordersCollected: { type: integer }
        collectionAmount: { type: number }
        routeSummary: { type: ["string", "null"] }

    # ── PAYROLL ───────────────────────────────────────
...