# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HRMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Visit {
  id: number;
  employeeId: number;
  /** @nullable */
  employeeName?: string | null;
  vendorId: number;
  /** @nullable */
  vendorName?: string | null;
  visitDate: string;
  /** @nullable */
  checkInTime?: string | null;
  /** @nullable */
  selfieUrl?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  remarks?: string | null;
  /** @nullable */
  meetingNotes?: string | null;
  /** @nullable */
  orderValue?: number | null;
  /** @nullable */
  nextFollowUp?: string | null;
  status: string;
  /** @nullable */
  invalidReason?: string | null;
  createdAt?: string;
}